pyproject.toml
[tool.poetry]
name = "qa-guru-diplom-mobile"
version = "0.1.0"
description = ""
authors = ["Rustam Dzhafarov <rustamd74@gmail.com>"]
readme = "README.md"
packages = [{include = "qa_guru_diplom_mobile"}]

[tool.poetry.dependencies]
python = "^3.10.5"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^3.2.1"
allure-pytest = "^2.13.1"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^1.0.0"


[tool.poetry.dev-dependencies]
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = " --alluredir=allure-results "
